+++ cmd/pint/lint.go
@@ -67,7 +67,7 @@ func actionLint(c *cli.Context) error {
 
 	minSeverity, err := checks.ParseSeverity(c.String(minSeverityFlag))
 	if err != nil {
-		return fmt.Errorf("invalid %s value: %w", minSeverityFlag, err)
+		return fmt.Errorf("invalid --%s value: %w", minSeverityFlag, err)
 	}
 
 	r := reporter.NewConsoleReporter(os.Stderr, minSeverity)